pdl: Parse structs like packets
The auto-generated “Packet” suffix now seems wrong, so I removed it.
This means that both “packet Foo” and “struct Foo” will generate the
following Rust types:

    pub struct Foo
    pub struct FooBuilder
    struct FooData

One potential problem with this: if someone has a “packet Packet”,
this now collides with “trait Packet” from our preamble. We might need
to correct for this when we integrate the generated code with clients.

Test: atest pdl_tests pdl_rust_generator_tests_{le,be}
Change-Id: I19354010e46ef1b5738fee9f7fb13ff7778d1b03