LeAudio: Fix timing issue on state machine reconfigure
There was a chance that during the consecutive reconfigurations,
state machine receives the previous configuration notification from
the ASE and considered it as part of current reconfiguration process.
It would skipp this ASE configuration which resulted in a mixed half-previous
half-new configuration (often asymmetric) that was a problem to some
devices.

Bug: 261417889
Bug: 229041682
Tag: #feature
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test --no-bazel-mode & tested manually
Change-Id: I43b6377ef9c1403b0a52dcd2a0a583caf6249b74