Revert "Revert^2 [Invisalign] Build changes to split out libblue..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78ce56e2-e278-46ce-b256-1bc0d0d71bfe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3d875bb30743c6e4437b4967efdbbbfd7f9e69f0