Add API to set connection interval at GATT connection
Adding a new BluetoothDevice#connectGatt API that takes a
connection priority constant as parameter as some users need
to be able to set the connection interval for a GATT
connection.
Adding new callbacks to receive updates for connection
priority.
Adding a new constant defining the connection interval for
CCC requirements.

Bug: 253544707
Tag: #feature
Test: CtsBluetoothMultiDevicesTestCases
Change-Id: I17282ef45d087b03ea6ae1fe20722de319361335