Revert "btm: Enhance with targeted announcements"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/249b4d6a-c0d9-423a-8fa1-9b8e05d3c8d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a86c2de90a1f3aa7ec632aaee5f1fe696a2b9d9