gatt: Fix handling MTU Exchange request
With this patch Android will always do only one MTU Exchange request
as per specification. For this reason Android always uses
GATT_MAX_MTU_SIZE as this is what GATT can easily handle in the
Bluetooth Stack.

This is done so the single application request cannot block other
applications from using bigger MTU than the one requested by the
first application.

Also, for backward compatibility, if MTU Exchange succeed and MTU turns
out to be bigger than what APP was requesting, then APP will be notified
with requested MTU.

Also, for backward compatibility, the for Data Length Extension the MTU
from the User is used and not the one which was internally requested by
the GATT

Bug: 257250188
Bug: 239776509
Test: atest BluetoothInstrumentationTests
Test: manual testing
Tag: #feature
Change-Id: I157d40040b53cd4595403d99e58e105709099684