[Non-Discoverable Mode] Hold ConnectComplete callback until local_address is known
If we get an incoming LE connection, we don't know the local_address
until the associated advertising set terminates. Rather than returning a
dummy value and *hoping* that we backfill in time, delay the event out
of GD until the value is ready.

Also fix the le_impl tests to enforce the new behavior, and add a flag.

Bug: 254314964
Test: compiles
Change-Id: I1d2d2f04a6c88c229a891705bf519b9ca5c69342