Revert "Add MessageTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d76c01f3-e93a-4127-920c-4fe9b879ec95). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I154756871800cd670e2f0dba15f69f34fbdab24f