Revert "Add MessageTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d03cefb-940c-4cf6-b62e-00c0c411dcc0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I16779652dfa0e126149b9d9d309d0b3ee4b50394