bta_gattc: Fix handling disconnect
Each of the p_clcb has the pointer to the server cache which keeps
number of clients, mtu, state etc.
Till now, state of the server was cleared when all p_clcb got
deallocated which usually happens on link disconnection.
However, if any of the client creates background connect, the p_clcb is
created and server would stay in the outdated state.
This leads to use invalid MTU after reconnection.

Bug: 263359541
Test: atest BluetoothInstrumenatationTest
Test: manual connect/disconnect GATT devices
Tag: #feature
Change-Id: I9d4a3b6d2c334af144e571025298fedea54025cd