Revert "RootCanal: Fix clang-tidy warnings"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e489012f-140e-481c-80d7-b9b8966bf95f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I797810b89e082962e0cc692065df70435b338f7d