Improve Implemetation of TRACE macros
The Problem: the source location (file, line num, func name))
contained in logs emitted by TRACE macros (e.g., BTIF_TRACE_ERROR)
does not point to TRACE statements, but the location of LOG_XXX
statements in LogMsg, which is not helpful for debugging.

Details: The TRACE macros expand to BT_TRACE which in turn expands to
a call to LogMsg. LogMsg uses LOG_XXX macros to emit logs.
The source location is added to the logs upon expansion of LOG_XXX
macros, thus it points to LOG_XXX statements in LogMsg function for
ALL TRACE statements, instead of the location of the
TRACE statements.

To make TRACE macros dump the source location of TRACE statements,
and make debugging easier, this patch changes the implementations
of BT_TRACE (which TRACE APIs ex) and LogMsg:
- LogMsg takes the source location (file, line num, func name) of
   the TRACE statements as additional args
- BT_TRACE passes the source location to LogMsg upon expansion

The LOG_XXX macros already dump source location. To avoid
duplication, a set of LOG_XXX_INT macros are introduced,
which does not add source location, and LOG_XXX is implemented
based in LOG_XXX_INT by adding the source location back.

Additionally this patch introduces a IGNORE_LOG_TAG_FILTER
macro for code where init log APIs (used by the log filtering)
are not available.

Bug: 264946508
Test: internal clean up, existing tests pass
Change-Id: I8d0816af48d1c3658421fc89620b60566618c296