Improve Implemetation of TRACE statements
The implenentation of TRACE statements (BTIF_TRACE_ERROR etc)
calls LogMsg, which in turn uses LOG macros to output log
statements. The source location info contained in the emitted
logs all point to the place in LogMsg where LOG macros are
used, instead of the location where the TRACE statements are
used.

This patch changes the signature of LogMsg to take the source
location of the TRACE statements and passes the
source location of TRACE statements to LogMsg.

Test: internal clean up, existing tests pass
Change-Id: I8d0816af48d1c3658421fc89620b60566618c296