Revert "Add tests to cover OPP activities tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a28fce87-77d6-4511-af89-a19555dc1833). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4caa074d2ed6419a0d3d3261dab39b94c5dda92e