Revert "Add tests to cover OPP activities tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9cdfbb1d-b318-4867-a4a9-b8a288869b53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idee1edf36e25571189df6a3db4f65141a5266033