Revert "Add tests to cover OPP activities tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/303a2de8-5afa-4d0d-a9cb-8514bb1dc505). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I878cf9f7757958572e0e15bd84f21b1b36f9ba16