Revert "Add tests to cover OPP activities tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03ac8a61-264a-414d-aa8a-17157ff01806). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I28138fcbefe3549efcd44989608af1d396fdf34c