Revert "Add tests to cover OPP activities tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5837fd86-e7b1-4e82-817f-c8ca65452d3b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7dfbd53addaddeed80023b1350e90f6940bab1f8