Revert "Add tests to cover OPP activities tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/421e0fac-05d9-4a1b-bf3c-7cbb9c4c8a35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f0779ca7a4c1ce717d0b5990be8259a537555d0