Revert "Add tests to cover OPP activities tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e14183e4-555b-43eb-b18f-05c06a7f4366). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a06b9e6bf8edd4e8384bb174e8000d6e376ff26