Revert "Add AdapterServiceBinderTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5afb296e-ac2c-4290-acfb-c97c984dc9cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7c6e4623af4b65cd84321e5cad6cc6379220a898