Rust PDL backend
This is a zero-copy Rust backend that supports most of PDL (excluding groups,
checksums, custom types), BUT only little-endian packets.

This is needed for Rust-GATT (scoped for U). It is not a permanent solution, I
expect to unify / replace this with the official Rust PDL once it is ready, but
until that point arrives I don't want to remain blocked.

Bug: 255880936
Test: unit

Change-Id: I7d2305c20288d576aa1a5731b0d7e54727366947