Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89795b6e-493d-4768-bc4a-bcc7d0406666). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I945e56a7f18156006dfaffcbc19f10c7657c00fa