Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8e3440dc-efda-4d0c-b32e-aeaffd4516c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie9a16e4c20843510b11812725cadc8852c6521ae