Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04a0bc21-d46e-42d4-9a6a-d11a28d613fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib83eee3ade1cd5858c40eb038f0d23c221e91f80