Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1c2208cc-1f3d-4bc7-ab7a-d61c51669fd1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id8f967d95b6d0b2490c89cf4cb76d148bb70e590