Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7adc108-4c53-45c0-9c22-6effc6f52ae6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9400b1fc2f10c922f97d0be446afe602cce8ebc5