Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/75ce7a81-bcd4-4f2a-a5ed-310b1629bf97). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I69c3150635b7c3fdc4dc5536792b39a00de90d8a