Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d742d75b-7c81-4b4d-b824-a1e83ec2f2bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I80e600a9bf82300a2b06aaa2011fedcf4671f700