Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e18ecc95-0b43-4fb8-8b9e-18e2b9fa7e06). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I239711ceda3df925739316003f0b0f0e47ba1b76