Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46aaeccb-f33d-404e-b4c1-3f9b86eea337). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe7fee26c929b963c56a0b98272b07b82324dd12