Revert "Add BluetoothHidDeviceBinderTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6d3d238-287c-4d43-8368-2bfdc83ca3f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6695cc19214127aec31d3e17f0e97c6fbd4b00b3