Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dcfe49f9-c579-4e5d-a1fb-fe0b442a80d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b27403a7795ba7ce3de2f4b62ad0c4afd95aa35