Revert "Add HfpNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b266aa5d-635e-48b3-b5b1-3b12d8f31c3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10b3da0ebc35a9b63d946ead51e3d51512251622