Revert "bta_gattc: Fix handling disconnect"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/116b714a-f94d-4d70-b190-fe2cc70e2179). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I04b0d9dd0a44a049a0c6bdf2f1d18d8bde7c4041