Revert "Make dependencies explicit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a90d1b09-0086-4636-9b98-477db119a38d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea51e87bd246cb0963ea967a342e4f067f7d8d20