Revert "Make dependencies explicit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6ddc122-d380-4a33-9c76-1fcbd9f35692). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idba20537f8c1c3737e7c3ce0b82f2fe143bc6a08