Floss: Don't spend time resolving names for unresponsive peers
Some faulty peers never replied to remote name resolving requests.
This could be a time-waster if we keep trying the RNR process on
every scanning cycle on these devices.

This CL introduces a cache to keep track such devices so we don't
spend time doing RNR on these devices for 5 minutes.

Bug: 236225499
Tag: #floss
Test: Use this trick to simulate unresponsive peer device: Put the
      peer device into pairing mode, then power off the peer when
      the DUT is just about to do RNR.
      Verify the peer is only RNR'd once every 5 minutes.

Change-Id: I6e4639a8df2957ad0156f6ec1ffeb9d846e3af72