Floss: End remote name resolving session if it takes too long
It takes about 5 seconds before we give up resolving names for one
peer device. If there is a lot of unresponsive peers in the
environment, it might take minutes for us just to wait all of them
to time out. During this time, we cannot scan, so the user might to
have to wait for a very long time before the target peer device is
discovered.

This CL introduces a timeout to end the resolving session even when
there are still devices need RNR.

Bug: 236225499
Tag: #floss
Test: Manual test, verify scanning is resumed early even when the
      environment has lots of unresponsive devices.

Change-Id: I0e21a66412bc45cabc0e852f0b9e524a3b0dff25