[Non-Discoverable Mode] Remove AdvertiserConfig
We have AdvertiserConfig and ExtendedAdvertiserConfig, but only use the
latter. This CL merges the two structs.

Note that there are still two *protos* with these two names, but the
protos are actually used in cert tests. But I don't see value in keeping
the separation in the actual implementation, since the reason for two
protos seems to just be convenience.

Test: gd cert
Bug: 254314964
Change-Id: Ibe58cfd44e124ea698932077da5f0b25de60fdbb