Revert "Make close part of the BluetoothProfile interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/490fce8d-8301-4221-acc8-1b4bae8c25eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I908c1125cc40d092c9939ee04f9278b169def098