Revert "Reenable net_test_stack_rfcomm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8849fb9-e629-4001-b043-9371d2b93e57). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d768bdca4b47658fbf4ea8bb560fa5012df17f1