Revert "Make close part of the BluetoothProfile interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50d9b0ac-9d1c-4fbf-a09f-7786266ab9c7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6631b356546195d7fe8deafb5c2eb323e8711d5