Revert "Make close part of the BluetoothProfile interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef4d6fd4-ee86-4495-98f5-c6bb3bf24377). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifba2f7f3ba81491002db80b0ecc59a5c18b0f17f