Revert "RootCanal: Enable bugprone-*, cert-* clang-tidy warnings"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/33598e2b-436c-48b0-a0a9-ee2c1bd5e7f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8054c1bab3b1d34e958fed4669eb3b714a085212