Revert "Make dependencies explicit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f079a0ee-5384-4ada-af05-ca71fe70ecd4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d940a3012a8c9cce70d8aa511d97169cb23369c