Revert "Add more tests for com.android.bluetooth package"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1c04edca-1151-4abd-aa88-e52723f01931). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3d6f7fef9d62166c721d2299340408151ee2f2b2