Revert "Add more tests for com.android.bluetooth package"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43b1117f-5201-43bb-a9bb-c6b61366e57e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic8a57b9f0fd266e6e00388f4bfaa45b24db90458