Revert "Prepare for clap 4."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4e6d2432-b703-44c0-a135-adf135f8ced0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic262440428e464238a7fb1c34195268fa566739a