device_iot_config: add init flag
An init flag can be toggled at any time so module methods
are not using it. The module is initialized regardless
if the flag is set. Only the public API is protected
and if the flag is not set then the calls to the API will return
immediately.

Bug: 261319328
Tag: #feature
Test: atest bluetooth_test_gd
BYPASS_LONG_LINES_REASON: Bluetooth likes 120 lines

Change-Id: I83541dccf7c24678a37ff4afc92a23fcfd14c203