vc: Dont ignore VC devices connected by another gatt client
Like in other services we should not remove disconnected
devices from the list of known and valid VC devices. Once
the device gets connected by another gatt client, we still
get the GATT connection event in VC, and if we ignore that
(which is the case when we remove the device from the list of known VC
devices), we will not be able to reconnect, since calling
BTA_GATTC_Open() will not trigger the connection event for
the already connected device.

Bug: 254810994
Tag: #feature
Test: atest --host bluetooth_vc_test --no-bazel-mode
Test: atest BluetoothInstrumentationTests
Change-Id: I46fa8e1da1b5c630fd633f51d29c0084367adb33