gd/hci: handle MSFT_LE_MONITOR_DEVICE_EVENT in gd layer
The LeScanningManager module should rely on the MsftExtensionManager
module so that it can use the SetScanningCallback function of the
latter. When the MsftExtensionManager module receives the Microsoft
vendor specific event MSFT_LE_MONITOR_DEVICE_EVENT, it can call the
OnTrackAdvFoundLost scanning callback to pass the event to higher
layers.

Bug: 261482382
Tag: #floss
Test: ./build.py --target test
Test: Manual test on a ChromeOS device, e.g., zork/morphius
      localhost # btclient
      bluetooth> le-scan register-scanner
      bluetooth> le-scan start-scan 1
      # If a target LE device is found and then lost, the log shows
      # something like
      bluetooth> btclient:info: Scan result lost: ScanResult {
        name: ...,
        address: ...,
        ...
      }

Change-Id: Ie255ef5c7d058a621e83c0eed241916de29a51a8