Revert "Fix isMetadataSynced_whenSynced_returnsTrue failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5748f198-5197-487a-bbbf-5a4b346dc119). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I987a066ba77b9d6754500b97d13d0d4a6fb8c469