Revert "Fix isMetadataSynced_whenSynced_returnsTrue failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff12f916-ad64-4e38-87c6-ae6e092d3d78). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I64601002f9f3e852ffd573d7dfa629f795cb9e8d