Revert "Fix isMetadataSynced_whenSynced_returnsTrue failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4468684-9422-49ee-ada0-c74150db6fa5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7fb7b6cd4d44bb62ed861632815c1754607104c5