Initialize remote_credit_count when L2CAP on connection request.
Without this change, remote_credit_count starts from zero and
overflows to 0xffff when the first packet is received. This is not a
big deal with the default Android credit params (only off by one) but
it leads to peer running out of credit if a low
L2CAP_LE_CREDIT_DEFAULT is used on the build.

Bug: 264642648
Bug: 264878187
Test: manual - verified L2CAP connection is not stuck when initial
credits is 0x18.

Change-Id: Ie630883ce70463dc18fc2a0c1c8b1ccc39cd2f1d