Revert "Fix isMetadataSynced_whenSynced_returnsTrue failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cecbbc01-a393-4f70-9847-b431fb82be28). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I50312952867bfab50e4b225ff56e972e8055ad57