Revert "Make dependencies explicit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/321efa48-c7f4-4245-9d57-f49ab2b1be36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie2f0940d466bd43db84cfece0b1a914b8895b6c0