gatt: Fix gatt reconnection
This is a regression after

70b63e4596 Ensure address removed from filter accept list

gatt/connection_manager is responsible for adding or removing device
from the allow list, based on the knownledge about number of interested
clients for the device.

It has been observed, that if single application disconnects non
connected device, it will be removed from the Allow list even there
might be others interested in the connection. This patch fixes that.

Bug: 266049504
Test: manual testing
Tag: #feature
Change-Id: Iaec02012f2766d290ce8fc2cb3bbf73436841f32