RootCanal: Declare the supported commands in controller properties
The Supported Commands mask is no longer set by DualModeController as
the list of implemented commands. This allows for setting unimplemented
commands as supported for feature discovery when testing new profiles.

The Supported Command mask value is unmodified in this CL, only adding
some commands that were implemented but missing from the HCI command
mapping:

- HCI Set Event Mask Page 2
- HCI Add SCO Connection
- HCI LE Remote Connection Parameter Request Reply
- HCI LE Remote Connection Parameter Request Negative Reply

The ability to configure supported commands from the JSON config file
is temporarily disabled as unimplemented commands now raise FATAL
errors and configured values are likely to be incorrect...

Test: m root-canal && atest pts-bot
Bug: 266087760
Change-Id: I9c8960ba65d885f3606a89d2dcc0781aa599c71f