Revert "Fix a test in BluetoothOppLauncherActivityTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/240db9a5-3b67-4382-b906-22681b7380a8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c022510693a167175d6a06c379540cd85b37f5e