Revert "leaudio: Clear group when devices got disconnected"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4eddaba8-b97e-4973-9fb4-9ae53dcb4e2e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If931f25f1b425e8580c1e07f17fdcadfbba17cef