Revert "leaudio: Clear group when devices got disconnected"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bc84891d-3720-4404-bb64-1a4e6d7e8c5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88e356ea59e1569f9e646dc7e6858b1d398613f1