Revert "enable scanRecord#getServiceData() for Scan"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21652360-abdd-4eb8-bf86-e6fe019992e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79e99358b75bb13101ddc8b5cb3d7e2542758235