Revert "enable scanRecord#getServiceData() for Scan"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/040983cf-39aa-4450-a7e9-d9387807e909). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7bba7cb6914de7e4f05692a688689d120ac9475d