Use correct address type based on address policy
There is an address policy global to the stack, as well as an address
type passed down from Java. We do not currently handle the various edge
cases caused by combinations of the two states properly in all the
advertising modes.

This CL pulls all that logic into a single function and reuses it across
advertising modes. The behavioral changes are to respect the
own_address_type in non-extended advertising, and to use the correct address
type when doing extended advertising with privacy disabled.

Test: unit
Bug: 268112598
Change-Id: I73f0fb18b6d32d708b081b8afa4e10122a1810c9