Add API for DCK connection interval
Add BluetoothGatt#CONNECTION_PRIORITY_DCK value to be used to set
connection priority to a value best supporting the Digital Car Key
implementation.

Note that the exact values defined will likely change, as it is
recommended to provide a range to the controller rather than a single
value for min and max.

Bug: 267229740
Test: Builds (functional changes to be added in follow-up CL, this is
adding the API and config value only)
Tag: #feature

Change-Id: Iff02e69caad079a627ff04ae6d593f58c897e16d