Revert "Fix coverity scan issue: uninitialized scalar variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/01b3adaa-ccef-4cf5-9d08-476a466e4da8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idc7d732b76782503104d4b90b6593453b15076c2