Revert "[redact_log] Mark the change in BluetoothDevice#toString"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b716d65a-cbe0-4e16-8e1d-44b481116a7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3fdee7b9f8c7bbe77bf6c537d5921edbac8a28a4