Revert "[redact_log]Address test failures caused by changes to B..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b716d65a-cbe0-4e16-8e1d-44b481116a7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1f0de08d29beb2d7cdf69c7a5152cbc3c9ced3c