Revert "[redact_log]Address test failures caused by changes to B..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0250fa5-184b-4ab4-8ca0-63ec75d3e66c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I43fd10ef3e9c772601b096ab8423737320c7b97b