Revert "bluetooth: if a scan failed to be resumed, don't remove"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efaa0a8f-8d29-48c7-930a-0f7e64ae1563). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ed9ffac10f5fe4040d1914a7ebcd4fc2bb11452