Revert "Adding new System API to set BT HCI snoop logging mode."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aa5aa3a2-3d8f-4a40-b864-c7c1a68417fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If8571d25fa5519357bc82e5e4fc66f2c0db69251