Floss: Add dynamically control HIDP and HOGP
This adds support to be able to dynamically activate and/or deactivate
HIDP and HOGP in the HID host module.

The profile is deactivated by exiting early in the OPEN request
function.

This will be first used by Floss admin policy feature. Without this CL,
we are not able to disallow HIDP but allow HOGP or the other way around.

Bug: 263755727
Test: tested it with the following CL.
Tag: #floss
Change-Id: I013304f50bcf1944f4f0c28fc38425cc638d0d90