LeAudio: Do not reconfigure to UNSPECIFIED scenario
When a remote does not declare support for  contexts like
NOTIFICATIONS, ALERTS, EMERGENCYALARM, INSTRUCTIONAL,
SOUNDEFFECTS and we would like to play them, we usually end
up using UNSPECIFIED context. In such case it is ok to use the
existing configuration and do not reconfigure for HQ playback
only configuration.

Bug: 269394021
Tag: #feature
Test: atest --host bluetooth_le_audio_client_test bluetooth_le_audio_test --no-bazel-mode
Change-Id: I050f310b50e6b1560a4b8bfc67aca170730f7c7d