Seen feature to add seen flag for message notifications
The goal is to provide support to apps such as Messenger to differentiate between new and pre-existing messages. The Seen flag will be set for all messages downloaded to the database during the initial import from the paired device. This is used to indicate that no new message notifications should be sent for messages that already existed on the paired device prior to connection with the HU. Thereafter, any message that is received by the paired device and imported into the HU database will not have the Seen flag set. The Seen flag will only be set for new messages after a message notification has been sent on the HU.

Tag: #feature
Bug: 234174225
Test: atest MapClientContentTest MapClientStateMachineTest
Change-Id: I4cb02788fa9201f5f8c7122003d21b9cbe49d82c