GTBS: Track inband ringtone status per device
With this patch, GTBS starts tracking inaband ringtone status per
device.

This is in order to expose proper enable state for the devices which are
Active in the system and not.

Also it might be used for the use case when, remote device disables
Ringtone from the available context types

Bug: 242685105
Bug: 260660659
Test: atest BluetoothInstrumentationTests
Tag: #feature
Change-Id: I2f13a25f64f3bf7b37a3218ad6040e9a5c439c13