Floss: Get device type from GD storage
Device type from the inquiry database could be misleading as a dual mode
device can be classified as LE. This happens during the initial pairing.
The inquiry database will be cleared when a new discovery is started and
if the LE advertisement from a dual mode device arrives, the device will
be considered as LE, even the LE advertisement is not connectable.

Test: scan and pair airpods from UI
Bug: 269521515

Change-Id: Ide6160c42d8ed5b6d8c6d8cdb4077184bf493bcd