Revert "pdl: Add support for struct fields"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e810f277-5cba-41c7-a7e0-f68bcf6ce9ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie91839e543f52f0d5236e9ac73c830378778166d