Revert "pdl: Add support for struct fields"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/da7bf239-cdc1-446e-943a-e8c668c2cfed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4c352c4355360eacb31d7d0cd8c2dba861854bd