Revert "pdl: Add support for struct fields"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1bfb27b-97bc-472e-895e-f8d654e88e5e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I472210cb922655a315d55d974bb6f64163a1c2fc