floss: Refactor le reconnection during suspend
During system suspend, we need to use a separate set of LE scan window
and intervals that are specifically tuned for this purpose. In order to
do this only when suspended, we add |SetSystemSuspendState| to indicate
when we're entering and existing suspend.

We set the suspend state via |AllowWakeByHid| and remove the suspend
state via |RestoreFilterAcceptList|. Currently, we only restore LE hid
devices on resume to the accept list but we should also restore
non-direct gatt connections in the future.

Bug: 271159571
Tag: #floss
Test: Manual test, ChromeOS Autotest and unit tests
Change-Id: I3e67f06f7667dec69379468cc51b8405cf10de10