Only initiate Codec Negotation if locally supported.
Currently, because of aosp/683369 the device will initiate
Codec Negotiation even if the remote device does not send the
required AT+BAC command during Service Level Connection setup.

This enables in the following out-of-spec sequence:
1. Device sends Codec Negotiation NOT supported to a Headset which does.
2. Headset respects the feature not supported, and does not send AT+BAC.
3. Device initiates Codec Negotiation anyway.

This check ensures that Device will not attempt to initiate
Codec Negotiation at all if the local feature is not enabled.

Bug: 267171863
Test: make
Change-Id: I446e071e3b66875a2bacde228841ca703acd5739