Remove descriptors unbonded device
When ACTION_LE_AUDIO_CONNECTION_STATE_CHANGED
is sent before ACTION_BOND_STATE_CHANGED is sent,
The state machine is removed but the device descriptor
is not removed from the hash map.

Which could results in NPE in dumpsys.

We could repro this by forgetting LE audio device
(not 100% but high chance).

This CL ensures the device descriptor is removed when
the state machine is removed.

Tag: #feature
Bug: 269969529
Test: manually forgetting LE devices and see dumpsys
and atest BluetoothInstrumentationTests

Change-Id: I0f9afe8ebb1712dd03b4072fe6f42a37919bb3c1