Revert "Bluetooth Quality Report feature flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/034d5f14-fb0c-47de-9074-8f4c75161515). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73c5695025781f167000a55173bb90f374c6271b