Revert "Bluetooth Quality Report feature flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1ab8dfa-9c7b-429a-a22f-c340db78160c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I74b3c1d31d0c0e5fa14ebf49d068f77f71e0466c