Revert "Bluetooth Quality Report features"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1ab8dfa-9c7b-429a-a22f-c340db78160c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I260ab1d4e3c5ec956ad2d2644d727d59fcfc8380