pdl: Generate simpler code for PDL structs
Before, we treated packets and structs the same: for a packet or
struct Foo, we would generate three Rust structs:

  struct Foo
  struct FooData
  struct FooBuilder

This doesn’t match the old bluetooth_packetgen compiler: it would only
generate a single Rust struct with public fields in case of a PDL
struct.

We now do the same for the new compiler. We reuse most of the
generator code from the old data struct generator since this is the
one that most closely resemble the struct we want.

Tag: #feature
Bug: 230475725
Test: atest pdl_tests pdl_rust_generator_tests_{le,be}
Change-Id: Idf2ae9cc1347d2f6b1a5b9b7c29f14e225e7080c