Do not reset battery level when HFP is disconnected
The battery level was reset when HFP was disconnected.

It was okay when HFP was the only way to get the battery level
but we are getting the battery level from the
battery service (BAS) as well so we need to
check if we should reset or not.

This change checks the condition and does not reset the
battery level if it is still valid information.

It also adds tests to ensure the functionality.

Bug: 269969529
Test: atest BluetoothInstrumentationTests and manually tested with LE HS and HFP
Change-Id: Iac16347045cac4201f0cd2132e51336159f03fed