Revert "Register metadata listener when bonding"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30780892-982c-4aec-98c3-5dfb7b0eb3bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99344c42e5e7fba866c629365b27647386d3a311