Revert "Introduce Bluetooth companion device manager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30780892-982c-4aec-98c3-5dfb7b0eb3bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifabdc4371c3523402d919cf22457af8f1a9de5bb