Revert "Address API feedback for Bluetooth Distance Measurement"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c710ddcb-2d07-47f6-8c6c-89a0e2a2894a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3b25c975888b35a51db6a40e2cb51fdf821112df