bluetooth: Fix connection manager issues
Promote gatt/connection_manager to manage direct connections as it
manage gatt users which requested it.
Without that, if more clients requested direct connection, and one of
them cancel it, we would end up with device being incorrectly
removed from allow list.

Note, with this change, direct_connections_ set is not used anymore in
acl_manager/le_impl.h as this is again handle in
gatt/connection_manager.cc

This is considered as a temporary fix, and proper one shall be handled
in b/272572974

Bug: 278199753
Test: manual
Tag: #feature

Change-Id: Iee744001f3a2f2193a500c52b573ddcc0eb7ff82