[Connection Manager] Remove locally_initiated field for LE
As per spec: The HCI_LE_Enhanced_Connection_Complete event indicates to
both of the Hosts forming the connection that a new connection has been
created. Upon the creation of the connection a Connection_Handle shall
be assigned by the Controller, and passed to the Host in this event. If
the connection creation fails, this event shall be provided to the Host
that had issued the HCI_LE_Create_- Connection or
HCI_LE_Extended_Create_Connection command.

Test: all existing
Bug: 272572974
Change-Id: I7359acbb907e42097633105d04f5fc5cb76ecf27