Floss: Fix a2dp source cleanup
In the a2dp source cleanup path, we deregistered BTA_IF_AV in bta_sys,
then send a disable event to the bta_sys. The second operation always
fail because the callback has been deregistered.

This fixes the issue by calling bta_av_disable directly given the fact
that we always call this function regardless the state we are at.

Bug: 264745185
Test: run AVRCP/CT/PTH/BV-01-C along with the top of this series
Tag: #floss

Change-Id: If882834c9ba2847f4ec630ede6aba468c28ce7fc