Floss: Add sysprop to configure gatt encryption after reconnecting.
Within the BTIF layer, the GATT automatically encrypts the link if
the device is bonded and then reconnected. However, for signed write
commands, the GATT must be in an unencrypted state. To accommodate
this, we have added a sysprop that can be configured to not encrypt
the link after reconnection.

Bug: 264727379
Test: Manually
Tag: #floss
Change-Id: I4005c241633d7e681a8fe7f88f95f54cf8e41218