Relan: Provide correct disconnection error code
As per spec:
> When the Controller receives the HCI_Disconnect command, it shall
send the HCI_Command_Status event to the Host. The
HCI_Disconnection_Complete event will occur at each Host when the
termination of the connection has completed, and on the local Host
also indicates that this command has been completed. The Reason
event parameter in the event on the local Host shall be set to the
value Connection Terminated by Local Host (0x16), while that on the
remote Host shall be set to the value of the Reason command
parameter.

This reverts commit dd1a1ba77de25dae18e761b049fbc8c7d4e445b7.

Test: atest pts-bot:HID/HOS/HCE/BV-03-I
Bug: 270629675
Change-Id: I33b71ceb96ea797ff4f7ed35f75496f0be7674f5