pdl: Ensure that all generated files compile
This adds a new unit test which will ensure that all files under
tests/generated/ actually compile.

We do this by combining all files into a gigantic Rust file. We then
compile this Rust file and that’s the actual test. The file has an
empty test to ensure that you see something useful if you do this by
hand via ‘cargo test’. The script is external from the Android.bp file
for the same reason: making it easy to run for a Cargo based workflow.

The compilation only takes a few seconds even though we’re compiling a
file with 13k lines of Rust.

Tag: #feature
Bug: 228306436
Test: atest pdl_tests pdl_rust_generator_tests_{le,be}
Change-Id: I4b213e99ff8a77801aa4f79f741add25176c9ff4