pdl: Use the analyzer ast for the default rust backend
The rust backend previously used the parser ast for the generation,
along with the legacy linter scope. This change migrates the generator
code to the analyzer ast while still using the legacy linter scope.

This will enable further code removal in the legacy linter as
the analyzer already inlines the group declarations.

Test: cargo test
Change-Id: Ic205b6a1452cd15d2dfd50c578c751a0245ee63f