Re-register callbacks when service is connected
When a callback is registered to
BluetoothHapClient, BluetoothLeAudio and BluetoothLeBroadcast,
and their corresponding service is not available,
they register callbacks when the service is available.

Before this change, they did when BT is up, which could be
before the service is available and it failed.

This change fixes the timing by listening to the service connection.
It also removes unnecessary broadcast receivers.

Tag: #feature
Bug: 267461828
Test: atest BluetoothInstrumentationTests
Change-Id: I12f49241b49fc2f793e7c2a25c66629aacd4c53f