Remove insufficient use of `rustfmt`
The 1.68.0 rust toolchain produces a `rustfmt` with new library
dependencies for which the current prebuilts setup does not work. The
current mechanism used for `rustfmt` here has prebuilt binaries being
copied, which is unnecessary, and we should instead have official build
rules for invoking those binaries from their prebuilt location, where
library dependencies can be satisfied. In order to unblock the
toolchain, the least invasive change is to disable using `rustfmt`
during the build for this generated code.

Bug: 270756468
Test: m rust
Change-Id: Id5c6400e1cafa462c7a2dd8700a4b9dd8e82b410