Revert "RootCanal: Implement the HCI Read Rssi command"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e97174e8-4cdb-4093-a0b8-d0dc9dbf9137). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2375ee219435e72ce83fa3bf4946b1036c7b7dd2