pdl: use prettyplease instead of rustfmt
The prettyplease library has been explicitly designed to be used for
programmatic formatting of Rust code. It is faster and simpler to use
than calling ‘rustfmt’.

Because ‘prettyplease’ sometimes uses a slightly different formatting
compared to ‘rustfmt’, I’m adding a ‘#![rustfmt::skip]’ like to the
top of every auto-generated file.

Tag: #refactor
Bug: 274187738
Test: atest pdl_tests pdl_rust_generator_tests_{le,be}
Change-Id: Ifd74db6c6bd06b020960d586ca803a26a4603c2d