Swap AVRCP queue Metadata with MediaController Metadata
When retrieving the Media metadata using
MediaSession#QueueItem, we can loose some information such
as media duration. To prevent this, we need to swap the
the stored QueueItem metadata with the data retrieved
from the MediaController, for the current playing media.

Now that the other media metadata is available for current
media QueueItem, we also need to update the equals overload
of the Metadata class to reflect these changes.

Bug: 235637427
Tag: #feature
Test: manual
Change-Id: Iefc7de8767ffab8fa14f9cf1f120ef4804791cd1