Revert "Create a new BluetoothSocketException class which is a s..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8faa180c-f795-4a0a-bd6d-f51516776ca1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I64c6ac9e3144090cca1c816481e2c373eed70ef2