Revert "Replace instances of IOException(msg) with BluetoothSock..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8faa180c-f795-4a0a-bd6d-f51516776ca1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc34480ab32facceba829464ea3fb5b5e7c7b22d