tbs: Don't create Bluetooth LE Call control without In Call service
This CL guards registration of Bluetooth LE Call control. Without this
guard there is possibility that BluetoothManagerService may broadcast an
onBluetoothStateChange intent that triggers binding InCallService. If
Bluetooth state change will be received before In Call service instance
will be created it may lead to double brearer registration and in fact
another registration from the same bearer would be rejected. If bearer
would not be registrated a call can't be set as active.

Tag: #feature
Bug: 273622512
Test: atest GoogleBluetoothInstrumentationTests
Change-Id: I0d6555d1e2a5d1859c87813f0e5ca4069ea7c8cb