Revert "Fix coverity scan issue: uninitialized scalar variable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7eec665-edf7-4ce5-b023-617dcf8c6bfc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb86e50768200ca404213c4f126b670fb93f2fc8