Revert "Revert "Add template to prevent sharding of some bt test..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ded7151-6e39-46c8-8991-bc56f7638273). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7c5bd1541e61f1ad30381f103418d68360d1291