Revert "Run GTests with ASAN on 32bit only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ded7151-6e39-46c8-8991-bc56f7638273). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0bf8f4a4782139790f1c99adc125a39b7f42b42