Revert "Run GTests with ASAN on 32bit only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9160688d-ce2d-4680-ad31-c721a2e3f9e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If94791a6d18ffcd1560e287fef372bdb52c858a8