Revert "Revert "Add template to prevent sharding of some bt test..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9160688d-ce2d-4680-ad31-c721a2e3f9e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If61259d044ddb6c6e235a0d0e709666b80fcf795