Revert "Revert "Add template to prevent sharding of some bt test..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a2811b2-fe5d-4e5e-ad03-d0b36780b01d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6aa1e482424171c38ad3ffaee5ca8562e32d6852