Revert "Run GTests with ASAN on 32bit only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a2811b2-fe5d-4e5e-ad03-d0b36780b01d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9b499f84e2dc0f33498794ddace48d5fa872d91c