Revert "PDL: support Erroneous Data Reporting"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d89756f-b133-4f27-8596-1c6c3a9feac1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I236dfead43719ceb7ced920d4ebe2aae55dfa79a