Revert "gd/hci: enable erroneous data reporting in controller"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85d5ecb8-8d51-401d-a1d4-467b95d7999a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I452da46a417c14c9e6112726cfbd2dce94cdb6fe