Revert "PDL: support Erroneous Data Reporting"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85d5ecb8-8d51-401d-a1d4-467b95d7999a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I192dedefe3d06e8049e7e8667ccd7f5b83122518