Revert "Revert "Add template to prevent sharding of some bt test..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bbc5cd00-2a98-41bf-85f7-cb44d92b595d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia33181e2d60d48cd8a2e8a3a1242e28d72120f41