Revert "Run GTests with ASAN on 32bit only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bbc5cd00-2a98-41bf-85f7-cb44d92b595d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I11dc2e925b92e63cbca059ca10b8bd2e59fe4686