Revert "gd/hci: enable erroneous data reporting in controller"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e86c134-714b-4dd8-9326-fc0a10420f5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7316befc1b4c652313f55d154bbb86d5c44ddc61