Revert "PDL: support Erroneous Data Reporting"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e86c134-714b-4dd8-9326-fc0a10420f5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec7b74de32bdaa5c77c10858283be2ba3d935d2f