Revert "Run GTests with ASAN on 32bit only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccc94c84-85a8-4445-a1b4-19d2d83d0063). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia34f0afe6f4cb3f6a97ec1e686bef239db104861