Revert "Revert "Add template to prevent sharding of some bt test..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccc94c84-85a8-4445-a1b4-19d2d83d0063). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id495472accac095cfaea659be72fb2c4224a8939