Revert "Use proper API and switch/case for bta search/discovery ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ddc5927-8cd3-4d28-9b0e-5f77a5252e9c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3945b380820854ab7772f00bc2bc8626b21f788b